Raise error when the same keyword is being used repeatedly
The usefulness of warnings is in decline: this change makes get_keyval() raise
an error condition instead of a warning (current behavior) when the
corresponding keyword is detected multiple times.  In nearly all cases, it is
clear from the documentation that only one of the values will be used.
One exception to this rule are the atom selection keywords, where some
(e.g. atomNameResidueRange) can be repeated but others (e.g. indexGroup,
atomsFile) cannot.

Coming improvements to the group class will allow all selection keywords to be
repeated.  This change is meant to prevent other instances of undefined
behavior that have not yet been considered.